+++ MyPlayground.playground/Pages/Edges.xcplaygroundpage/timeline.xctimeline
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Timeline
+   version = "3.0">
+   <TimelineItems>
+      <LoggerValueHistoryTimelineItem
+         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/MyPlayground.playground/Pages/Edges.xcplaygroundpage#CharacterRangeLen=96&amp;CharacterRangeLoc=1238&amp;EndingColumnNumber=0&amp;EndingLineNumber=30&amp;StartingColumnNumber=1&amp;StartingLineNumber=29&amp;Timestamp=552207559.433742"
+         selectedRepresentationIndex = "0"
+         shouldTrackSuperviewWidth = "NO">
+      </LoggerValueHistoryTimelineItem>
+   </TimelineItems>
+</Timeline>
