+++ MyPlayground.playground/Pages/Tiny Constraints.xcplaygroundpage/Contents.swift
@@ -9,8 +9,9 @@ import TinyConstraints
 
 //: First we set up our ViewController and the UIView that we want to add the constraints to.
 let viewController = UIViewController()
+PlaygroundPage.current.liveView = viewController
 let constraintView = UIView()
-let ding = TinyTheme.backgroundColor
+let ding = TinyTheme.constraintViewColor
 viewController.view.backgroundColor = TinyTheme.backgroundColor
 constraintView.backgroundColor = TinyTheme.constraintViewColor
 //: Always remember to add the view as a subview before you add constraints, otherwise your app will crash 💥
@@ -29,13 +30,6 @@ constraintView.height(100)
 constraintView.width(100)
 
 //: or even shorter, both at once:
-constraintView.size(CGSize(width: 100, height: 100))
-
-//: Center
-//: Constraining the center of a view to its superview with NSLayoutConstraint:
-
-//: with TinyConstraints:
+//constraintView.size(CGSize(width: 100, height: 100))
 
 //: [Edges](@next)
-
-PlaygroundPage.current.liveView = viewController
