+++ MyPlayground.playground/contents.xcplayground
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<playground version='6.0' target-platform='ios' executeOnSourceChanges='false'>
+    <pages>
+        <page name='Tiny Constraints'/>
+        <page name='Edges'/>
+    </pages>
+</playground>
