--- TinyConstraints.playground/Pages/0 - Tiny Constraints.xcplaygroundpage/timeline.xctimeline
@@ -1,51 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Timeline
-   version = "3.0">
-   <TimelineItems>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=0&amp;CharacterRangeLoc=1208&amp;EndingColumnNumber=19&amp;EndingLineNumber=36&amp;StartingColumnNumber=19&amp;StartingLineNumber=36&amp;Timestamp=552154481.147296"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=1&amp;CharacterRangeLoc=957&amp;EndingColumnNumber=4&amp;EndingLineNumber=27&amp;StartingColumnNumber=3&amp;StartingLineNumber=27&amp;Timestamp=552154481.1475199"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=1&amp;CharacterRangeLoc=957&amp;EndingColumnNumber=4&amp;EndingLineNumber=27&amp;StartingColumnNumber=3&amp;StartingLineNumber=27&amp;Timestamp=552154481.1477"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=51&amp;CharacterRangeLoc=880&amp;EndingColumnNumber=4&amp;EndingLineNumber=25&amp;StartingColumnNumber=35&amp;StartingLineNumber=23&amp;Timestamp=552154481.1478651"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=15&amp;CharacterRangeLoc=1193&amp;EndingColumnNumber=19&amp;EndingLineNumber=36&amp;StartingColumnNumber=4&amp;StartingLineNumber=36&amp;Timestamp=552154481.148026"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=1&amp;CharacterRangeLoc=957&amp;EndingColumnNumber=4&amp;EndingLineNumber=27&amp;StartingColumnNumber=3&amp;StartingLineNumber=27&amp;Timestamp=552154481.148187"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=0&amp;CharacterRangeLoc=1208&amp;EndingColumnNumber=19&amp;EndingLineNumber=36&amp;StartingColumnNumber=19&amp;StartingLineNumber=36&amp;Timestamp=552154481.148351"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=0&amp;CharacterRangeLoc=1208&amp;EndingColumnNumber=19&amp;EndingLineNumber=36&amp;StartingColumnNumber=19&amp;StartingLineNumber=36&amp;Timestamp=552154481.148508"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/0%20-%20Tiny%20Constraints.xcplaygroundpage#CharacterRangeLen=52&amp;CharacterRangeLoc=1044&amp;EndingColumnNumber=0&amp;EndingLineNumber=32&amp;StartingColumnNumber=1&amp;StartingLineNumber=31&amp;Timestamp=552154481.148672"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-   </TimelineItems>
-</Timeline>
