--- TinyConstraints.playground/Pages/1 - Animating Constraints.xcplaygroundpage/Contents.swift
@@ -1,7 +1,0 @@
-//: [Previous](@previous)
-
-import Foundation
-
-var str = "Hello, playground"
-
-//: [Next](@next)
