--- TinyConstraints.playground/Pages/1 - Edges.xcplaygroundpage/timeline.xctimeline
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Timeline
-   version = "3.0">
-   <TimelineItems>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/1%20-%20Edges.xcplaygroundpage#CharacterRangeLen=96&amp;CharacterRangeLoc=1250&amp;EndingColumnNumber=0&amp;EndingLineNumber=31&amp;StartingColumnNumber=2&amp;StartingLineNumber=29&amp;Timestamp=552200545.600033"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/1%20-%20Edges.xcplaygroundpage#CharacterRangeLen=96&amp;CharacterRangeLoc=1249&amp;EndingColumnNumber=0&amp;EndingLineNumber=30&amp;StartingColumnNumber=1&amp;StartingLineNumber=29&amp;Timestamp=552200545.600202"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-      <LoggerValueHistoryTimelineItem
-         documentLocation = "file:///Users/marijnschilling/Projects/TinyConstraints/TinyConstraints.playground/Pages/1%20-%20Edges.xcplaygroundpage#CharacterRangeLen=0&amp;CharacterRangeLoc=1395&amp;EndingColumnNumber=0&amp;EndingLineNumber=32&amp;StartingColumnNumber=0&amp;StartingLineNumber=32&amp;Timestamp=552200545.60032"
-         selectedRepresentationIndex = "0"
-         shouldTrackSuperviewWidth = "NO">
-      </LoggerValueHistoryTimelineItem>
-   </TimelineItems>
-</Timeline>
