--- TinyConstraints.playground/contents.xcplayground
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<playground version='6.0' target-platform='ios' display-mode='rendered'>
-    <pages>
-        <page name='0 - Tiny Constraints'/>
-        <page name='1 - Edges'/>
-        <page name='1 - Animating Constraints'/>
-    </pages>
-</playground>
